Refactor print to logging; PunishmentPattern in its own module
